+++ src/algorithms/fov.rs
@@ -1,18 +1,62 @@
-use crate::{DiagonalDirection, Direction, Hex};
+use crate::{Direction, Hex};
 use std::collections::HashSet;
 
-pub fn range_fov(coord: Hex, range: u32, visible: impl Fn(Hex) -> bool) -> HashSet<Hex> {
+/// Computes a field of view around `coord` in a given `range`.
+/// This algorithm takes in account coordinates *visibility* through the `blocking` argument.
+/// (*Blocking* coordinates should return `true`)
+///
+/// # Examples
+///
+/// - Compute field of view with no boundaries and some blocking tiles
+///
+/// ```rust
+/// # use hexx::*;
+/// # use std::collections::HashSet;
+/// use hexx::algorithms::range_fov;
+///
+/// let pos = hex(0, 0);
+/// let range = 10;
+/// let blocking_coords: HashSet<Hex> = HashSet::new();
+/// // Add blocking coordinates
+/// // blocking_coords.insert(hex(2, 0));
+/// // ..
+/// let fov = range_fov(pos, range, |h| blocking_coords.contains(&h));
+/// ```
+pub fn range_fov(coord: Hex, range: u32, blocking: impl Fn(Hex) -> bool) -> HashSet<Hex> {
     coord
         .ring(range)
-        .flat_map(|target| coord.line_to(target).take_while(|h| visible(*h)))
+        .flat_map(|target| coord.line_to(target).take_while(|h| !blocking(*h)))
         .collect()
 }
 
-pub fn dual_fov(
+/// Computes a field of view around `coord` in a given `range` towards `direction` with 120 degrees
+/// vision.
+/// This algorithm takes in account coordinates *visibility* through the `blocking` argument.
+/// (*Blocking* coordinates should return `true`)
+///
+/// # Examples
+///
+/// - Compute drectional field of view with no boundaries and some blocking tiles
+///
+/// ```rust
+/// # use hexx::*;
+/// # use std::collections::HashSet;
+/// use hexx::algorithms::directional_fov;
+///
+/// let pos = hex(0, 0);
+/// let range = 10;
+/// let dir = Direction::Top;
+/// let blocking_coords: HashSet<Hex> = HashSet::new();
+/// // Add blocking coordinates
+/// // blocking_coords.insert(hex(2, 0));
+/// // ..
+/// let fov = directional_fov(pos, range, dir, |h| blocking_coords.contains(&h));
+/// ```
+pub fn directional_fov(
     coord: Hex,
     range: u32,
-    visible: impl Fn(Hex) -> bool,
     direction: Direction,
+    blocking: impl Fn(Hex) -> bool,
 ) -> HashSet<Hex> {
     let [a, b] = [direction.diagonal_left(), direction.diagonal_right()];
     coord
@@ -21,32 +65,6 @@ pub fn dual_fov(
             let way = coord.diagonal_way_to(*h);
             way == a || way == b
         })
-        .flat_map(|target| coord.line_to(target).take_while(|h| visible(*h)))
-        .collect()
-}
-
-pub fn directional_fov(
-    coord: Hex,
-    range: u32,
-    visible: impl Fn(Hex) -> bool,
-    direction: Direction,
-) -> HashSet<Hex> {
-    coord
-        .ring(range)
-        .filter(|h| coord.way_to(*h) == direction)
-        .flat_map(|target| coord.line_to(target).take_while(|h| visible(*h)))
-        .collect()
-}
-
-pub fn diagonal_fov(
-    coord: Hex,
-    range: u32,
-    visible: impl Fn(Hex) -> bool,
-    direction: DiagonalDirection,
-) -> HashSet<Hex> {
-    coord
-        .ring(range)
-        .filter(|h| coord.diagonal_way_to(*h) == direction)
-        .flat_map(|target| coord.line_to(target).take_while(|h| visible(*h)))
+        .flat_map(|target| coord.line_to(target).take_while(|h| !blocking(*h)))
         .collect()
 }