+++ src/algorithms/mod.rs
@@ -1,5 +1,5 @@
 mod fov;
 mod pathfinding;
 
-pub use fov::{diagonal_fov, directional_fov, dual_fov, range_fov};
+pub use fov::{directional_fov, range_fov};
 pub use pathfinding::a_star;